cosmetics and minor refactoring before merge
