+++ ecfeed/src/com/testify/ecfeed/ui/editor/ModelEditor.java
@@ -35,9 +35,11 @@ import org.eclipse.jdt.core.IPackageFragmentRoot;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jface.dialogs.ErrorDialog;
 import org.eclipse.jface.dialogs.MessageDialog;
+import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.jface.window.Window;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.IEditorInput;
+import org.eclipse.ui.IPersistableElement;
 import org.eclipse.ui.PartInitException;
 import org.eclipse.ui.dialogs.SaveAsDialog;
 import org.eclipse.ui.forms.editor.FormEditor;
@@ -52,18 +54,52 @@ import com.testify.ecfeed.serialization.IModelSerializer;
 import com.testify.ecfeed.serialization.ParserException;
 import com.testify.ecfeed.serialization.ect.EctParser;
 import com.testify.ecfeed.serialization.ect.EctSerializer;
+import com.testify.ecfeed.ui.common.Constants;
 import com.testify.ecfeed.ui.common.IFileInfoProvider;
+import com.testify.ecfeed.ui.common.Messages;
 
 public class ModelEditor extends FormEditor implements IFileInfoProvider{
 
-	private static final int MAX_SUBTREE_SIZE = 10000;	
 	private RootNode fModel;
 	private ModelPage fModelPage;
 	private ModelOperationManager fModelManager;
 	private ObjectUndoContext fUndoContext;
-	private ModelXmlViewerTextEditor fXmlViewerPageEditor;
-	private int fXmlViewerPageEditorIndex = -1;
-	
+	private ModelSourceEditor fSourcePageEditor;
+	private int fSourcePageIndex = -1;
+
+	public class SourceEditorInput implements IEditorInput{
+
+		@Override
+		@SuppressWarnings("rawtypes")
+		public Object getAdapter(Class adapter) {
+			return null;
+		}
+
+		@Override
+		public boolean exists() {
+			return true;
+		}
+
+		@Override
+		public ImageDescriptor getImageDescriptor() {
+			return null;
+		}
+
+		@Override
+		public String getName() {
+			return "XML";
+		}
+
+		@Override
+		public IPersistableElement getPersistable() {
+			return null;
+		}
+
+		@Override
+		public String getToolTipText() {
+			return "XML view of model";
+		}
+	}
 
 	private class ResourceChangeReporter implements IResourceChangeListener {
 		@Override
@@ -144,48 +180,47 @@ public class ModelEditor extends FormEditor implements IFileInfoProvider{
 		}
 		return root;
 	}
-	
+
 	@Override
 	protected void addPages() {
 		try {
 			setPartName(getEditorInput().getName());
 			addPage(fModelPage = new ModelPage(this));
-			
-			addXmlViewerPage();
+
+			addSourcePage();
 
 		} catch (PartInitException e) {
 			ErrorDialog.openError(getSite().getShell(),
-					"Error creating nested text editor",
+					"Error while creating nested text editor",
 					null, e.getStatus());
 		}
 	}
-	
-	private void addXmlViewerPage() throws PartInitException {
-		IEditorInput editorInput = new ModelXmlViewerPageInput();
+
+	private void addSourcePage() throws PartInitException {
+		IEditorInput editorInput = new SourceEditorInput();
 		setPartName(getEditorInput().getName());
-		fXmlViewerPageEditor = new ModelXmlViewerTextEditor(getSite().getShell());
-		
-		fXmlViewerPageEditorIndex = addPage(fXmlViewerPageEditor, editorInput);
-		setPageText(fXmlViewerPageEditorIndex, fXmlViewerPageEditor.getTitle());
+		fSourcePageEditor = new ModelSourceEditor(getSite().getShell());
+
+		fSourcePageIndex = addPage(fSourcePageEditor, editorInput);
+		setPageText(fSourcePageIndex, fSourcePageEditor.getTitle());
 	}
-	
+
 	@Override
 	protected void pageChange(int newPageIndex) {
 		super.pageChange(newPageIndex);
-		
-		if (newPageIndex != fXmlViewerPageEditorIndex)
+
+		if (newPageIndex != fSourcePageIndex)
 			return;
-		
-		if (fModel.subtreeSize() <= MAX_SUBTREE_SIZE) {
-			refreshXmlViewerPageWithSerializedModel();
+
+		if (fModel.subtreeSize() <= Constants.SOURCE_VIEWER_MAX_SUBTREE_SIZE) {
+			refreshSourceViewer();
 		}
 		else {
-			fXmlViewerPageEditor.refreshContent("\n  Warning. Model exceeds " 
-					+ MAX_SUBTREE_SIZE + " items. It can not be displayed." );
+			fSourcePageEditor.refreshContent(Messages.MODEL_SOURCE_SIZE_EXCEEDED(Constants.SOURCE_VIEWER_MAX_SUBTREE_SIZE));
 		}
 	}
-	
-	private void refreshXmlViewerPageWithSerializedModel()
+
+	private void refreshSourceViewer()
 	{
 		ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
 		IModelSerializer serializer = new EctSerializer(outputStream);
@@ -196,11 +231,11 @@ public class ModelEditor extends FormEditor implements IFileInfoProvider{
 		catch(Exception e){
 			MessageDialog.openError(Display.getCurrent().getActiveShell(),
 					"Error", "Could not serialize the file:" + e.getMessage());
-		}			
-		
-		fXmlViewerPageEditor.refreshContent(outputStream.toString());
+		}
+
+		fSourcePageEditor.refreshContent(outputStream.toString());
 	}
-	
+
 	@Override
 	public void doSave(IProgressMonitor monitor) {
 		IFile file = ((FileEditorInput)getEditorInput()).getFile();
