--- ecfeed/src/com/testify/ecfeed/ui/editor/ModelXmlViewerPageInput.java
@@ -1,39 +1,0 @@
-package com.testify.ecfeed.ui.editor;
-
-import org.eclipse.jface.resource.ImageDescriptor;
-import org.eclipse.ui.IEditorInput;
-import org.eclipse.ui.IPersistableElement;
-
-public class ModelXmlViewerPageInput implements IEditorInput{
-
-	@Override
-	@SuppressWarnings("rawtypes")
-	public Object getAdapter(Class adapter) {
-		return null;
-	}
-
-	@Override
-	public boolean exists() {
-		return true;
-	}
-
-	@Override
-	public ImageDescriptor getImageDescriptor() {
-		return null;
-	}
-
-	@Override
-	public String getName() {
-		return "XML";
-	}
-
-	@Override
-	public IPersistableElement getPersistable() {
-		return null;
-	}
-
-	@Override
-	public String getToolTipText() {
-		return "XML view of model";
-	}
-}
